refactor(86c21k8e5): rename files to match plural convention
