--- src/promotions/entity/promotion.entity.ts
@@ -1,32 +1,0 @@
-import { Entity, Column, PrimaryGeneratedColumn, CreateDateColumn, UpdateDateColumn } from 'typeorm';
-
-@Entity('promotion')
-export class Promotion {
-  @PrimaryGeneratedColumn('uuid')
-  uuid: string;
-
-  @Column({ length: 50 })
-  name: string;
-
-  @Column({ name: 'is_active', type: 'boolean' })
-  isActive: boolean;
-
-  @Column({ name: 'start_date', type: 'date' })
-  startDate: Date;
-
-  @Column({ name: 'end_date', type: 'date' })
-  endDate: Date;
-
-  @CreateDateColumn({
-    name: 'created_at',
-    type: 'timestamp'
-  })
-  createdAt: Date;
-
-  @UpdateDateColumn({
-    name: 'updated_at',
-    type: 'timestamp',
-    nullable: true
-  })
-  updatedAt: Date;
-}
